change document structure. Gamesection and hostSection staged inside GameField component
