+++ src/Navbar/navbar.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+function Navbar() {
+	return(
+		<navbar>
+			<ul>
+				<li>Вход</li>
+				<li>Иконка пользователя с входом в профиль</li>
+			</ul>
+		</navbar>
+	)
+}
+
+
+
+export default Navbar
