--- src/host/HostSection.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import Avatar from "../components/Avatar";
-
-function HostSection() {
-    return (
-        <div className="host-section">
-            < Avatar />
-            <div className="info">
-                <p>info info info info</p>
-            </div>
-        </div>
-    )
-}
-
-export default HostSection;
