+++ src/index.css
@@ -19,10 +19,15 @@ code {
 }
 
 .wrapper {
-  display: flex;
+	display: flex;
+	flex-direction: column;
   width: 100vw;
   height: 100vh;
   background-color: blue;
+}
+
+.GAME_FIELD_WRAPPER{
+	display: flex;
 }
 
 .host-section {
